fix(products) se cambia userId por JWT, para obtener el id del usuario

+++ donatello/src/main/java/com/igrowker/donatello/services/impl/ProductServiceImpl.java
@@ -2,16 +2,21 @@ package com.igrowker.donatello.services.impl;
 
 import com.igrowker.donatello.dtos.ProductDTO;
 import com.igrowker.donatello.exceptions.ConflictException;
+import com.igrowker.donatello.exceptions.ForbiddenException;
 import com.igrowker.donatello.exceptions.NotFoundException;
 import com.igrowker.donatello.mappers.ProductMapper;
+import com.igrowker.donatello.models.Customer;
 import com.igrowker.donatello.models.Product;
 import com.igrowker.donatello.repositories.IProductRepository;
 import com.igrowker.donatello.repositories.IUserRepository;
+import com.igrowker.donatello.services.IAuthService;
 import com.igrowker.donatello.services.IProductService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductServiceImpl implements IProductService {
@@ -24,6 +29,8 @@ public class ProductServiceImpl implements IProductService {
 
     @Autowired
     private ProductMapper productMapper;
+    @Autowired
+    IAuthService authService;
 
     @Override
     public List<ProductDTO> getProducts() {
@@ -31,7 +38,10 @@ public class ProductServiceImpl implements IProductService {
     }
 
     @Override
-    public ProductDTO add(ProductDTO productDto) {
+    public ProductDTO add(HttpHeaders headers, ProductDTO productDto) {
+        Integer userId = authService.getLoguedUser(headers).getId();
+        productDto.setUserID(userId);
+
         Boolean existsProduct = productRepository.existsByName(productDto.getName());
         if (!existsProduct) {
             Product product = productMapper.toProduct(productDto);
@@ -43,18 +53,30 @@ public class ProductServiceImpl implements IProductService {
     }
 
     @Override
-    public ProductDTO update(Integer productId, ProductDTO productDto) {
+    public ProductDTO update(HttpHeaders headers,Integer productId, ProductDTO productDto) {
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new NotFoundException("Product not found!"));
+
+        Integer userId = authService.getLoguedUser(headers).getId();
+        if(! product.getCustomUser().getId().equals(userId)){
+            throw new ForbiddenException("The user cannot update someone else's product.");
+        }
+
         product.setCustomUser(userRepository.getReferenceById(productDto.getUserID()));
         productMapper.updateProduct(product, productDto);
         return productMapper.toProductDto(productRepository.save(product));
     }
 
     @Override
-    public ProductDTO delete(Integer productId) {
+    public ProductDTO delete(HttpHeaders headers,Integer productId) {
         Product product = productRepository.findById(productId)
                 .orElseThrow(() -> new NotFoundException("Product not found!"));
+
+        Integer userId = authService.getLoguedUser(headers).getId();
+        if(! product.getCustomUser().getId().equals(userId)){
+            throw new ForbiddenException("The user cannot delete someone else's product.");
+        }
+
         ProductDTO productDto = productMapper.toProductDto(product);
         productRepository.deleteById(productId);
         return productDto;
